requirements.txt
geog==0.0.2
numpy==1.16.2
python-telegram-bot==12.5.1
Shapely==1.6.4.post2
